factor out util.exists_in_cmr function
